+++ component/icon/UserDarkIcon.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+
+const UserDarkIcon = () => {
+  return (
+    <svg
+      width="16"
+      height="18"
+      viewBox="0 0 16 18"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M8 0.75C7.00544 0.75 6.05161 1.14509 5.34835 1.84835C4.64509 2.55161 4.25 3.50544 4.25 4.5C4.25 5.49456 4.64509 6.44839 5.34835 7.15165C6.05161 7.85491 7.00544 8.25 8 8.25C8.99456 8.25 9.94839 7.85491 10.6517 7.15165C11.3549 6.44839 11.75 5.49456 11.75 4.5C11.75 3.50544 11.3549 2.55161 10.6517 1.84835C9.94839 1.14509 8.99456 0.75 8 0.75ZM4 10.25C3.00544 10.25 2.05161 10.6451 1.34835 11.3483C0.645088 12.0516 0.25 13.0054 0.25 14V15.188C0.25 15.942 0.796 16.584 1.54 16.705C5.818 17.404 10.182 17.404 14.46 16.705C14.8201 16.6464 15.1476 16.4616 15.3839 16.1837C15.6202 15.9057 15.75 15.5528 15.75 15.188V14C15.75 13.0054 15.3549 12.0516 14.6517 11.3483C13.9484 10.6451 12.9946 10.25 12 10.25H11.66C11.475 10.25 11.291 10.28 11.116 10.336L10.25 10.619C8.78796 11.0962 7.21204 11.0962 5.75 10.619L4.884 10.336C4.70865 10.2789 4.5254 10.2499 4.341 10.25H4Z"
+        fill="#31273F"
+      />
+    </svg>
+  );
+};
+
+export default UserDarkIcon;
