--- src/app/bottomsheet/page.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import TheBottomSheet from "../../../component/bottomSheet/BottomSheet";
-
-const page = () => {
-  return (
-    <>
-      <TheBottomSheet />
-    </>
-  );
-};
-
-export default page;
