+++ utils/hooks/usePostUserDiet.ts
@@ -1,6 +1,6 @@
 import { useMutation } from "@tanstack/react-query";
 import { postUserDiet } from "../api/AxiosSetting";
-import { UserPostDietData } from "@/app/main/page";
+import { UserPostDietData } from "@/app/maintest/page";
 import { useToast } from "@chakra-ui/react";
 import { useUser } from "./useUser";
